Splitting the mitigation experiments to two classes
